+++ Source/Injector/Pegasus.cs
@@ -15,7 +15,7 @@ using System.Windows.Forms;
 
 internal static class Program {
     internal const string DllName = "BedrockUtilityFramework.Xray.dll";
-    internal const string ExpectedHash = "9BFE3F8093EB55023246AD13B3D319BB85923594D410203A99DCE845C02529B0";
+    internal const string ExpectedHash = "EAAC7C2BC8D313ACCEA5EF2D3ECBAA8801D8D6C7B00C44297D0B5917F03C863A";
     internal static string DllPath { get { return Path.Combine(AppDomain.CurrentDomain.BaseDirectory, DllName); } }
     [STAThread] static int Main(string[] args) {
         if (args.Length == 1 && args[0] == "--test-host") { Thread.Sleep(30000); return 0; }